Move `LocatedPrunableTree` and its helper types into `prunable` submodule
